Topの画面をActivityに
