--- app/src/main/kotlin/com/numero/itube/activity/MainActivity.kt
@@ -1,56 +1,0 @@
-package com.numero.itube.activity
-
-import android.os.Bundle
-import android.view.Menu
-import android.view.MenuItem
-import androidx.appcompat.app.AppCompatActivity
-import com.numero.itube.R
-import com.numero.itube.extension.findFragment
-import com.numero.itube.extension.replace
-import com.numero.itube.fragment.FavoriteVideoListFragment
-import com.numero.itube.fragment.SettingsBottomSheetFragment
-import com.numero.itube.fragment.SettingsFragment
-import com.numero.itube.repository.db.FavoriteVideo
-import kotlinx.android.synthetic.main.activity_main.*
-
-class MainActivity : AppCompatActivity(),
-        FavoriteVideoListFragment.FavoriteFragmentListener,
-        SettingsFragment.SettingsFragmentListener {
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
-        setSupportActionBar(toolbar)
-
-        if (findFragment(R.id.container) == null) {
-            replace(R.id.container, FavoriteVideoListFragment.newInstance(), false)
-        }
-    }
-
-    override fun onCreateOptionsMenu(menu: Menu): Boolean {
-        menuInflater.inflate(R.menu.menu_main, menu)
-        return true
-    }
-
-    override fun onOptionsItemSelected(item: MenuItem): Boolean {
-        return when (item.itemId) {
-            R.id.action_settings -> {
-                SettingsBottomSheetFragment.newInstance().show(supportFragmentManager)
-                true
-            }
-            R.id.action_search -> {
-                startActivity(SearchActivity.createIntent(this))
-                true
-            }
-            else -> super.onOptionsItemSelected(item)
-        }
-    }
-
-    override fun showLicenses() {
-        startActivity(LicensesActivity.createIntent(this))
-    }
-
-    override fun showVideo(video: FavoriteVideo) {
-        startActivity(PlayerActivity.createIntent(this, video))
-    }
-}
