+++ app/src/main/kotlin/com/numero/itube/activity/TopActivity.kt
@@ -1,45 +1,38 @@
-package com.numero.itube.fragment
+package com.numero.itube.activity
 
-import android.content.Context
 import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
+import android.view.Menu
+import android.view.MenuItem
+import androidx.appcompat.app.AppCompatActivity
 import androidx.core.view.isInvisible
-import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProviders
 import androidx.recyclerview.widget.LinearLayoutManager
 import com.numero.itube.R
 import com.numero.itube.contract.FavoriteVideoListContract
 import com.numero.itube.extension.component
 import com.numero.itube.extension.observeNonNull
+import com.numero.itube.fragment.SettingsBottomSheetFragment
+import com.numero.itube.fragment.SettingsFragment
 import com.numero.itube.presenter.FavoriteVideoListPresenter
 import com.numero.itube.repository.FavoriteVideoRepository
-import com.numero.itube.repository.db.FavoriteVideo
 import com.numero.itube.view.adapter.FavoriteVideoListAdapter
 import com.numero.itube.viewmodel.FavoriteVideoListViewModel
-import kotlinx.android.synthetic.main.fragment_favorite_video_list.*
+import kotlinx.android.synthetic.main.activity_top.*
 import javax.inject.Inject
 
-class FavoriteVideoListFragment : Fragment() {
+class TopActivity : AppCompatActivity(), SettingsFragment.SettingsFragmentListener {
 
     @Inject
     lateinit var favoriteVideoRepository: FavoriteVideoRepository
 
     private lateinit var presenter: FavoriteVideoListContract.Presenter
     private val videoListAdapter: FavoriteVideoListAdapter = FavoriteVideoListAdapter()
-    private var listener: FavoriteFragmentListener? = null
-
-    override fun onAttach(context: Context?) {
-        super.onAttach(context)
-        if (context is FavoriteFragmentListener) {
-            listener = context
-        }
-    }
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         component?.inject(this)
+        setContentView(R.layout.activity_top)
+        setSupportActionBar(toolbar)
 
         val viewModel = ViewModelProviders.of(this).get(FavoriteVideoListViewModel::class.java)
         viewModel.videoList.observeNonNull(this) {
@@ -50,16 +43,9 @@ class FavoriteVideoListFragment : Fragment() {
         }
 
         presenter = FavoriteVideoListPresenter(viewModel, favoriteVideoRepository)
-    }
 
-    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
-        return inflater.inflate(R.layout.fragment_favorite_video_list, container, false)
-    }
-
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
         videoListAdapter.setOnItemClickListener {
-            listener?.showVideo(it)
+            startActivity(PlayerActivity.createIntent(this, it))
         }
         videoRecyclerView.apply {
             layoutManager = LinearLayoutManager(context)
@@ -68,16 +54,31 @@ class FavoriteVideoListFragment : Fragment() {
         }
     }
 
+    override fun onCreateOptionsMenu(menu: Menu): Boolean {
+        menuInflater.inflate(R.menu.menu_main, menu)
+        return true
+    }
+
+    override fun onOptionsItemSelected(item: MenuItem): Boolean {
+        return when (item.itemId) {
+            R.id.action_settings -> {
+                SettingsBottomSheetFragment.newInstance().show(supportFragmentManager)
+                true
+            }
+            R.id.action_search -> {
+                startActivity(SearchActivity.createIntent(this))
+                true
+            }
+            else -> super.onOptionsItemSelected(item)
+        }
+    }
+
     override fun onResume() {
         super.onResume()
         presenter.loadFavoriteVideoList()
     }
 
-    interface FavoriteFragmentListener {
-        fun showVideo(video: FavoriteVideo)
-    }
-
-    companion object {
-        fun newInstance(): FavoriteVideoListFragment = FavoriteVideoListFragment()
+    override fun showLicenses() {
+        startActivity(LicensesActivity.createIntent(this))
     }
 }
