+++ app/src/main/kotlin/com/numero/itube/di/ApplicationComponent.kt
@@ -2,8 +2,8 @@ package com.numero.itube.di
 
 import com.numero.itube.activity.PlayerActivity
 import com.numero.itube.activity.SearchActivity
+import com.numero.itube.activity.TopActivity
 import com.numero.itube.fragment.ChannelVideoListFragment
-import com.numero.itube.fragment.FavoriteVideoListFragment
 import dagger.Component
 import javax.inject.Singleton
 
@@ -15,9 +15,9 @@ import javax.inject.Singleton
     (RepositoryModule::class)
 ])
 interface ApplicationComponent {
-    fun inject(favoriteVideoListFragment: FavoriteVideoListFragment)
     fun inject(channelVideoListFragment: ChannelVideoListFragment)
 
+    fun inject(topActivity: TopActivity)
     fun inject(searchActivity: SearchActivity)
     fun inject(playerActivity: PlayerActivity)
 }
